CB-11259. Relocate CM based diagnostics flow.
