+++ core/src/main/java/com/sequenceiq/cloudbreak/core/flow2/cmdiagnostics/event/CmDiagnosticsCollectionFailureEvent.java
@@ -1,6 +1,6 @@
-package com.sequenceiq.cloudbreak.core.flow2.diagnostics.event;
+package com.sequenceiq.cloudbreak.core.flow2.cmdiagnostics.event;
 
-import static com.sequenceiq.cloudbreak.core.flow2.diagnostics.event.CmDiagnosticsCollectionStateSelectors.FAILED_CM_DIAGNOSTICS_COLLECTION_EVENT;
+import static com.sequenceiq.cloudbreak.core.flow2.cmdiagnostics.event.CmDiagnosticsCollectionStateSelectors.FAILED_CM_DIAGNOSTICS_COLLECTION_EVENT;
 
 import com.sequenceiq.cloudbreak.common.event.Selectable;
 import com.sequenceiq.common.model.diagnostics.CmDiagnosticsParameters;
