+++ core/src/main/java/com/sequenceiq/cloudbreak/core/flow2/cmdiagnostics/handler/CmDiagnosticsCleanupHandler.java
@@ -1,7 +1,7 @@
-package com.sequenceiq.cloudbreak.core.flow2.diagnostics.handler;
+package com.sequenceiq.cloudbreak.core.flow2.cmdiagnostics.handler;
 
-import static com.sequenceiq.cloudbreak.core.flow2.diagnostics.event.CmDiagnosticsCollectionHandlerSelectors.INIT_CM_DIAGNOSTICS_EVENT;
-import static com.sequenceiq.cloudbreak.core.flow2.diagnostics.event.CmDiagnosticsCollectionStateSelectors.START_CM_DIAGNOSTICS_COLLECTION_EVENT;
+import static com.sequenceiq.cloudbreak.core.flow2.cmdiagnostics.event.CmDiagnosticsCollectionHandlerSelectors.CLEANUP_CM_DIAGNOSTICS_EVENT;
+import static com.sequenceiq.cloudbreak.core.flow2.cmdiagnostics.event.CmDiagnosticsCollectionStateSelectors.FINISH_CM_DIAGNOSTICS_COLLECTION_EVENT;
 
 import java.util.HashSet;
 import java.util.Map;
@@ -13,8 +13,8 @@ import org.slf4j.LoggerFactory;
 import org.springframework.stereotype.Component;
 
 import com.sequenceiq.cloudbreak.core.flow2.diagnostics.DiagnosticsFlowService;
-import com.sequenceiq.cloudbreak.core.flow2.diagnostics.event.CmDiagnosticsCollectionEvent;
-import com.sequenceiq.cloudbreak.core.flow2.diagnostics.event.CmDiagnosticsCollectionFailureEvent;
+import com.sequenceiq.cloudbreak.core.flow2.cmdiagnostics.event.CmDiagnosticsCollectionEvent;
+import com.sequenceiq.cloudbreak.core.flow2.cmdiagnostics.event.CmDiagnosticsCollectionFailureEvent;
 import com.sequenceiq.common.api.telemetry.model.DiagnosticsDestination;
 import com.sequenceiq.common.model.diagnostics.CmDiagnosticsParameters;
 import com.sequenceiq.flow.reactor.api.event.EventSender;
@@ -24,9 +24,9 @@ import reactor.bus.Event;
 import reactor.bus.EventBus;
 
 @Component
-public class CmDiagnosticsInitHandler extends EventSenderAwareHandler<CmDiagnosticsCollectionEvent> {
+public class CmDiagnosticsCleanupHandler extends EventSenderAwareHandler<CmDiagnosticsCollectionEvent> {
 
-    private static final Logger LOGGER = LoggerFactory.getLogger(CmDiagnosticsInitHandler.class);
+    private static final Logger LOGGER = LoggerFactory.getLogger(CmDiagnosticsCleanupHandler.class);
 
     @Inject
     private EventBus eventBus;
@@ -34,13 +34,13 @@ public class CmDiagnosticsInitHandler extends EventSenderAwareHandler<CmDiagnost
     @Inject
     private DiagnosticsFlowService diagnosticsFlowService;
 
-    public CmDiagnosticsInitHandler(EventSender eventSender) {
+    public CmDiagnosticsCleanupHandler(EventSender eventSender) {
         super(eventSender);
     }
 
     @Override
     public String selector() {
-        return INIT_CM_DIAGNOSTICS_EVENT.selector();
+        return CLEANUP_CM_DIAGNOSTICS_EVENT.selector();
     }
 
     @Override
@@ -51,21 +51,21 @@ public class CmDiagnosticsInitHandler extends EventSenderAwareHandler<CmDiagnost
         CmDiagnosticsParameters parameters = data.getParameters();
         Map<String, Object> parameterMap = parameters.toMap();
         try {
-            LOGGER.debug("CM based diagnostics collection initialization started. resourceCrn: '{}', parameters: '{}'", resourceCrn, parameterMap);
+            LOGGER.debug("CM based diagnostics cleanup started. resourceCrn: '{}', parameters: '{}'", resourceCrn, parameterMap);
             if (DiagnosticsDestination.SUPPORT.equals(parameters.getDestination())) {
-                LOGGER.debug("CM based diagnostics uses SUPPORT destination, no support specific init step yet.");
+                LOGGER.debug("CM based diagnostics uses SUPPORT destination, no support specific cleanup step yet.");
             } else {
-                diagnosticsFlowService.init(resourceId, parameterMap, new HashSet<>());
+                diagnosticsFlowService.cleanup(resourceId, parameterMap, new HashSet<>());
             }
             CmDiagnosticsCollectionEvent diagnosticsCollectionEvent = CmDiagnosticsCollectionEvent.builder()
                     .withResourceCrn(resourceCrn)
                     .withResourceId(resourceId)
-                    .withSelector(START_CM_DIAGNOSTICS_COLLECTION_EVENT.selector())
+                    .withSelector(FINISH_CM_DIAGNOSTICS_COLLECTION_EVENT.selector())
                     .withParameters(parameters)
                     .build();
             eventSender().sendEvent(diagnosticsCollectionEvent, event.getHeaders());
         } catch (Exception e) {
-            LOGGER.debug("CM based diagnostics collection initialization failed. resourceCrn: '{}', parameters: '{}'.", resourceCrn, parameterMap, e);
+            LOGGER.debug("CM based diagnostics cleanup failed. resourceCrn: '{}', parameters: '{}'.", resourceCrn, parameterMap, e);
             CmDiagnosticsCollectionFailureEvent failureEvent = new CmDiagnosticsCollectionFailureEvent(resourceId, e, resourceCrn, parameters);
             eventBus.notify(failureEvent.selector(), new Event<>(event.getHeaders(), failureEvent));
         }
